+++ src/components/utilities/Defense/Overview/Accordion/DefenseOverviewAccordion.vue
@@ -1,9 +1,11 @@
 <template>
   <div class="accordion">
     <div class="row">
-      <div v-for="defense in defenses" :key="defense.incrementId" class="col-md-4">
-        <Defense :defense="defense" :collapsed="defense.userData?.isCollapsed" />
-      </div>
+      <slot name="defense-list" v-for="defense in defenses" :defense="defense" :key="defense.incrementId">
+        <div class="col-md-4">
+          <Defense :defense="defense" :collapsed="defense.userData?.isCollapsed" />
+        </div>
+      </slot>
     </div>
   </div>
 </template>